move `models.v2` to `v2.models`
