+++ chemprop/v2/models/models/classification.py
@@ -1,7 +1,7 @@
 from torch import Tensor, nn
 
 from chemprop.data.v2.dataloader import TrainingBatch
-from chemprop.models.v2.models.base import MPNN
+from chemprop.v2.models.models.base import MPNN
 
 
 class ClassificationMPNN(MPNN):
