+++ chemprop/v2/models/models/multiclass.py
@@ -2,8 +2,8 @@ from torch import Tensor
 from torch.nn import functional as F
 
 from chemprop.data.v2.dataloader import TrainingBatch
-from chemprop.models.v2.modules import MessagePassingBlock
-from chemprop.models.v2.models.base import MPNN
+from chemprop.v2.models.modules import MessagePassingBlock
+from chemprop.v2.models.models.base import MPNN
 
 
 class MulticlassMPNN(MPNN):
